add: add more content to ReactUseLocalStorage.tsx
I was able to understand and even modify the useLocalStorage hook from
react-use. My modification was able to eliminate initializing the
deserializer every time the hook ran at the expense of adding some extra
logic in the set function. Further, I was able to eliminate an eslint
warning regarding the dependency array for the useCallback that wraps
set.